Centralise 'renderData' and provide a function to access it.
