+++ Source/Plugins/Core/com.equella.core/js/tsrc/components/RichTextEditor.tsx
@@ -18,7 +18,7 @@
 import * as React from "react";
 import { Editor } from "@tinymce/tinymce-react";
 import { AxiosPromise, AxiosResponse } from "axios";
-import { AppConfig } from "../AppConfig";
+import { AppConfig, getRenderData } from "../AppConfig";
 
 import "tinymce/tinymce";
 
@@ -59,9 +59,8 @@ import "tinymce/plugins/toc";
 import "tinymce/plugins/visualblocks";
 import "tinymce/plugins/visualchars";
 import "tinymce/plugins/wordcount";
-import type { RenderData } from "../mainui";
 
-declare const renderData: RenderData | undefined;
+const renderData = getRenderData();
 
 // from https://github.com/tinymce/tinymce/blob/26b948ac85b75991ab9e50d0affdf4f5c0b34f65/modules/tinymce/src/core/main/ts/api/file/BlobCache.ts#L31-L39
 export interface BlobInfo {
